+++ src/app/builds/build-details/build-details.component.ts
@@ -18,6 +18,7 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
   listener: any = null;
   logEvents: MainLog[] = [];
   container: HTMLElement;
+  wasScrolledToBottom: boolean;
 
   constructor(
     private route: ActivatedRoute,
@@ -44,6 +45,7 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
       }
       if (!!window.EventSource && !this.listener) {
         this.source.addEventListener('message', (message: MessageEvent) => {
+          this.wasScrolledToBottom = this.isScrolledToBottom();
           // When it comes to SSE, the MessageEvent.data is always a string
           const data = JSON.parse(message.data);
           this.logEvents.push(data);
@@ -69,7 +71,7 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
   }
 
   private stayScrolledToBottom(): void {
-    if (this.isScrolledToBottom()) {
+    if (this.wasScrolledToBottom) {
       this.scrollToBottom();
     }
   }