+++ src/app/builds/build-details/build-details.component.ts
@@ -2,32 +2,38 @@ import { ConfigService } from './../../shared/config/config.service';
 import { AfterViewChecked, Component, OnDestroy, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { BuildService, MainLog } from 'api-client';
-import { WharfProject } from 'src/app/models/main-project.model';
 import { BuildStatus } from '../../models/build-status';
+import { Title } from '@angular/platform-browser';
+
+const enum Tabs {
+  Logs = 0,
+  Artifacts,
+}
 
 @Component({
   selector: 'wh-build-details',
   templateUrl: './build-details.component.html',
 })
 export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecked {
-  buildId: string;
-  project: WharfProject;
+  buildId: number;
   buildStatus?: BuildStatus;
   myData: any;
   source: EventSource;
   listener: any = null;
   logEvents: MainLog[] = [];
   container: HTMLElement;
+  activeTabIndex = 0;
   wasScrolledToBottom: boolean;
 
   constructor(
     private route: ActivatedRoute,
     private buildService: BuildService,
-    private configService: ConfigService) { }
+    private configService: ConfigService,
+    private titleService: Title) { }
 
   ngOnInit(): void {
-    this.buildId = this.route.snapshot.paramMap.get('buildId');
-    this.buildService.buildBuildidGet(Number(this.buildId)).subscribe(build => {
+    this.buildId = Number(this.route.snapshot.paramMap.get('buildId'));
+    this.buildService.buildBuildidGet(this.buildId).subscribe(build => {
       this.buildStatus = build.statusId;
       this.connect();
     });
@@ -60,7 +66,7 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
         console.warn(`Unknown build status ID '${this.buildStatus}' on build ID '${this.buildId}'.`);
       }
     }
-    this.buildService.buildBuildidLogGet(Number(this.buildId)).subscribe(log => {
+    this.buildService.buildBuildidLogGet(this.buildId).subscribe(log => {
       this.logEvents = log;
       this.logEvents.sort((a, b) => a.logId - b.logId);
     });
@@ -68,6 +74,10 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
 
   ngOnDestroy() {
     this.source?.removeEventListener('message', this.listener);
+  }
+
+  onTabChanged() {
+    this.updateTitle();
   }
 
   private stayScrolledToBottom(): void {
@@ -83,5 +93,13 @@ export class BuildDetailsComponent implements OnInit, OnDestroy, AfterViewChecke
   private isScrolledToBottom(): boolean {
     // The -2 is to have some margin for error
     return (window.innerHeight + window.pageYOffset) >= document.body.scrollHeight - 2;
+  }
+
+  private updateTitle() {
+    if (this.activeTabIndex === Tabs.Logs) {
+      this.titleService.setTitle(`Build ${this.buildId} - Wharf`);
+    } else if (this.activeTabIndex === Tabs.Artifacts) {
+      this.titleService.setTitle(`Artifacts - Build ${this.buildId} - Wharf`);
+    }
   }
 }