+++ CHANGELOG.md
@@ -17,6 +17,9 @@ This project tries to follow [SemVer 2.0.0](https://semver.org/).
 - Added functionality for favoriting/unfavoriting a project in the project
   details view. (#58)
 
+- Fixed page not scrolling down when new build logs arrive in bulk and the page
+  is scrolled to the bottom. (#60)
+
 ## v1.3.3 (2021-07-13)
 
 - Security fix by changing version of nginx base image in Dockerfile from